+++ .idea/dictionaries/zen.xml
@@ -8,6 +8,7 @@
       <w>geofenceables</w>
       <w>geofences</w>
       <w>groupable</w>
+      <w>parameterizer</w>
       <w>photoable</w>
       <w>photoables</w>
       <w>reacterable</w>
